pyproject.toml
[project]
name = "baby-style-kg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dj-rest-auth[with-social]>=7.2.0",
    "django-allauth>=65.15.1",
    "django-cors-headers>=4.9.0",
    "django-jazzmin>=3.0.4",
    "django-solo>=2.5.1",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=12.2.0",
    "python-decouple>=3.8",
]
